Tweaking blockquote :before/:after quotes.
The previous CSS rules were a bit simpler, but they could overlap the
blockquote text if the open/close quote charaters were too large (e.g.
if a different font was used).

This new code is a bit more convoluted, but will make the open/close
quote expand to the outside of the blockquote, so it will never overlap
the quoted text.