Adding Greasy Fork logo and profile link.
The logo contains some background text. I've decided to keep the text as
text (instead of path) for simplicity. Sure, the logo is not 100%
accurate, but it is close enough. Also, I don't know which font was used
in the original logo, and the original font was rendered at a specific
pixel size, which leads to hinting (i.e. aligning the lines with the
pixel grid), which is harder to replicate in SVG.

Also changing the calculation at aboutme-hex-horizontal.svg, so that the
liquid template will automatically adjust the width according to the
number of items. Requires jekyll 3.0 or later (because of `ceil`
liquid filter).