Fixing broken link to Brave web browser
I removed the link to the referral program in d009008d57963b3ab00306ff45f154b7ebb69be7, so this link has been broken for 10 months!

The fact is that Brave had TWO programs: the referral one to get people to download and install the browser, and the rewards one that gives BATs (Brave Attention Tokens) to creators. It took me a long time to figure it out.