Added Steam Library Filters
It's a very useful tool (and there are other nice tools on the same
website). It's more complete and more comprehensive than Steam's
built-in searching and filtering.

The logo is simple, could be redrawn from scratch, but I'm just going to
reuse the plain PNG because I don't want to bother re-creating an SVG
for it.